+++ app_plus/src/main/java/org/gkisalatiga/plus/screen/ScreenLiturgi.kt
@@ -245,17 +245,7 @@ class ScreenLiturgi : ComponentActivity() {
                                     maxLines = 3,
                                     overflow = TextOverflow.Ellipsis
                                 )
-                                // The file size.
-                                Row {
-                                    Icon(Icons.Default.Download, "File download size icon", modifier = Modifier.scale(0.8f).padding(end = 5.dp))
-                                    Text(
-                                        size,
-                                        fontSize = 14.sp,
-                                        fontWeight = FontWeight.Normal,
-                                        maxLines = 1,
-                                        overflow = TextOverflow.Ellipsis
-                                    )
-                                }
+
                                 // The downloaded PDF badge.
                                 val isDownloadedTitle = stringResource(R.string.pdf_already_downloaded_localized)
                                 val badgeColor = Color(Colors.MAIN_PDF_DOWNLOADED_BADGE_COLOR)
@@ -269,6 +259,18 @@ class ScreenLiturgi : ComponentActivity() {
                                             maxLines = 1,
                                             overflow = TextOverflow.Ellipsis,
                                             color = badgeColor
+                                        )
+                                    }
+                                } else {
+                                    // The file size.
+                                    Row {
+                                        Icon(Icons.Default.Download, "File download size icon", modifier = Modifier.scale(0.8f).padding(end = 5.dp))
+                                        Text(
+                                            size,
+                                            fontSize = 14.sp,
+                                            fontWeight = FontWeight.Normal,
+                                            maxLines = 1,
+                                            overflow = TextOverflow.Ellipsis
                                         )
                                     }
                                 }